Restructure the SCIP solver backend into multiple files
